+++ src/adapter/desktop.tsx
@@ -7,6 +7,9 @@ import { showNotification } from "@mantine/notifications";
 import { actions, store } from "~/store";
 import { SurrealistAdapter } from "./base";
 import { printLog } from '~/util/helpers';
+import { getActiveSurreal } from '~/surreal';
+import { map } from 'radash';
+import { TableSchema, TableField, TableIndex, TableEvent } from '~/typings';
 
 const WAIT_DURATION = 1000;
 
@@ -77,6 +80,67 @@ export class DesktopAdapter implements SurrealistAdapter {
 
 	async openUrl(url: string) {
 		open(url);
+	}
+
+	async fetchSchema() {
+		const surreal = getActiveSurreal();
+		const dbResponse = await surreal.query('INFO FOR DB');
+		const dbResult = dbResponse[0].result;
+
+		if (!dbResult) {
+			return [];
+		}
+
+		const databaseInfo = await map(Object.values(dbResult.tb), definition => {
+			return invoke<TableSchema>('extract_table_definition', { definition });
+		});
+
+		return map(databaseInfo, (async table => {
+			const tbResponse = await surreal.query(`INFO FOR TABLE ${table.name}`);
+			const tableInfo = tbResponse[0].result;
+
+			const fieldInfo = await map(Object.values(tableInfo.fd), definition => {
+				return invoke<TableField>('extract_field_definition', { definition });
+			});
+
+			const indexInfo = await map(Object.values(tableInfo.ix), definition => {
+				return invoke<TableIndex>('extract_index_definition', { definition });
+			});
+
+			const eventInfo = await map(Object.values(tableInfo.ev), definition => {
+				return invoke<TableEvent>('extract_event_definition', { definition });
+			});
+
+			const mappedFields = fieldInfo.map(field => {
+				let kind = field.kind;
+				let kindTables: string[] = [];
+				let kindGeometry: string[] = [];
+
+				if (field.kind.startsWith('record')) {
+					kindTables = field.kind.replace('record(', '').replace(')', '').split(',');
+					kind = 'record';
+				}
+
+				if (field.kind.startsWith('geometry')) {
+					kindGeometry = field.kind.replace('geometry(', '').replace(')', '').split(',');
+					kind = 'geometry';
+				}
+
+				return {
+					...field,
+					kind,
+					kindGeometry,
+					kindTables
+				};
+			});
+
+			return {
+				schema: table,
+				fields: mappedFields,
+				indexes: indexInfo,
+				events: eventInfo
+			};
+		}));
 	}
 
 	initDatabaseEvents() {